refactor: move governance clients
